requirements.txt
selenium==4.31.0
pytest
allure-pytest
webdriver-manager
python-dotenv
